creating user_repository package
