+++ lib/app/app.dart
@@ -1,0 +1,2 @@
+export 'router/router.dart';
+export 'view/app.dart';
