+++ lib/app/view/app.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:user_list/user_list/router/router.dart';
+import 'package:user_list/app/app.dart';
 import 'package:user_list/user_list/bloc/users_bloc.dart';
 import 'package:user_list/edit_user/bloc/user_bloc.dart';
