+++ lib/edit_user/bloc/user_bloc.dart
@@ -1,6 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:user_list/user_list/models/models.dart';
+import 'package:user_repository/user_repository.dart';
 
 part 'user_event.dart';
 part 'user_state.dart';
