+++ lib/main.dart
@@ -1,4 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:user_list/app.dart';
+import 'package:user_list/app/view/app.dart';
 
 void main() => runApp(App());
