+++ lib/user_list/bloc/users_bloc.dart
@@ -2,7 +2,7 @@ import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'dart:async';
 import 'package:meta/meta.dart';
-import 'package:user_list/user_list/models/models.dart';
+import 'package:user_repository/user_repository.dart';
 
 part 'users_event.dart';
 part 'users_state.dart';
@@ -16,8 +16,8 @@ class UsersBloc extends Bloc<UsersEvent, UsersState> {
       UpdateUsersList event, Emitter<UsersState> emit) async {
     emit(state.copyWith(status: UsersStatus.loading));
     try {
-      final _usersListService = UsersListService();
-      final List<User> users = await _usersListService.getUsersList();
+      final _usersListRepository = UsersListRepository();
+      final List<User> users = await _usersListRepository.getUsersList();
       return emit(state.copyWith(
         users: users,
         status: UsersStatus.success,
