--- lib/user_list/models/models.dart
@@ -1,2 +1,0 @@
-export 'user.dart';
-export 'users_list.dart';
