+++ packages/user_repository/lib/src/models/models.dart
@@ -1,0 +1,1 @@
+export 'user.dart';
