+++ packages/user_repository/lib/src/users_list_repository.dart
@@ -1,12 +1,11 @@
 import 'dart:async';
-import 'dart:io';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
-import 'package:user_list/user_list/models/user.dart';
+import 'models/models.dart';
+import 'user_repository.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:user_list/user_list/models/user.dart';
 
-class UsersListService {
+class UsersListRepository {
   Future<List<User>> getUsersList() async {
     // https://assessment-users-backend.herokuapp.com/users.json
     final uri =
@@ -15,7 +14,7 @@ class UsersListService {
     final jsonList = jsonDecode(response.body);
     print(jsonList[0]);
 
-    final _userService = UserService();
+    final _userService = UserRepository();
     List<User> usersList = [];
     for (var jsonUser in jsonList) {
       final user = await User.fromJson(jsonUser);
