+++ packages/user_repository/lib/user_repository.dart
@@ -1,1 +1,5 @@
+library user_repository;
 
+export 'src/models/models.dart';
+export 'src/user_repository.dart';
+export 'src/users_list_repository.dart';
