+++ packages/user_repository/pubspec.yaml
@@ -1,0 +1,21 @@
+name: user_repository
+description: A dart repository which manages the user domain.
+
+publish_to: 'none' 
+version: 1.0.0+1
+
+environment:
+  sdk: ">=2.15.1 <3.0.0"
+
+dependencies:
+  equatable: ^2.0.3
+  http: ^0.13.4
+  path_provider: ^2.0.8
+  intl: ^0.17.0
+
+dev_dependencies:
+  bloc_test: ^9.0.2
+  flutter_lints: ^1.0.4
+  flutter_test:
+    sdk: flutter
+  mocktail: ^0.2.0
